split core editor file;
