+++ packages/ui-schema/src/Schema/Editor.js
@@ -1,158 +1,6 @@
 import React from 'react';
-import {List} from "immutable";
 import {SchemaEditorProvider, useSchemaEditor} from "./EditorStore";
-import ErrorBoundary from "react-error-boundary";
-
-const MyFallbackComponent = ({componentStack, error}) => (
-    <div>
-        <p><strong>Oops! An error occured!</strong></p>
-        <p>Here’s what we know…</p>
-        <p><strong>Error:</strong> {error.toString()}</p>
-        <p><strong>Stacktrace:</strong> {componentStack}</p>
-    </div>
-);
-
-const getPlugin = (current, widgetStack) => {
-    return current < widgetStack.length ? widgetStack[current] : false;
-};
-
-const WidgetStackRenderer = ({current, Widget, widgetStack, ...props}) => {
-    // plugin layer
-
-    // first Plugin
-    const Plugin = getPlugin(current, widgetStack);
-
-    return <ErrorBoundary FallbackComponent={MyFallbackComponent}>
-        {Plugin ? <Plugin {...props} Widget={Widget} valid errors={List()} widgetStack={widgetStack} current={current}/> : 'plugin-stack-error'}
-    </ErrorBoundary>;
-};
-
-const FinalWidgetRenderer = ({Widget, ...props}) => {
-    return <Widget {...props}/>;
-};
-
-const NextPluginRenderer = ({current, Widget, widgetStack, ...props}) => {
-    const next = current + 1;
-    const Plugin = getPlugin(next, widgetStack);
-
-    return next < widgetStack.length ?
-        Plugin ? <Plugin {...props} Widget={Widget} widgetStack={widgetStack} current={next}/> : 'plugin-error' :
-        <FinalWidgetRenderer {...props} Widget={Widget}/>;
-};
-
-const DumpWidgetRenderer = React.memo(({Widget, widgetStack: widgetStack, ...props}) => {
-    // abstraction layer that only pushes the to-rendering's values
-    // thus no hook inside widget needed, widget and widget-plugins can be pure/dump for its scope
-
-    return <ErrorBoundary FallbackComponent={MyFallbackComponent}>
-        {widgetStack ?
-            <WidgetStackRenderer {...props} Widget={Widget} widgetStack={widgetStack} current={0}/> :
-            <FinalWidgetRenderer {...props} Widget={Widget}/>}
-    </ErrorBoundary>;
-});
-
-const SchemaWidgetRenderer = ({schema, parentSchema, storeKeys, showValidity: showValidityOverride, ...props}) => {
-    const {store, setData, widgets, showValidity, validity, onValidity, t} = useSchemaEditor();
-    const {widgetStack} = widgets;
-    const type = schema.get('type');
-    const widget_name = schema.get('widget');
-
-    // getting the to-render component based on if it finds a custom object-widget or a widget extending native-types
-    let Widget = null;
-
-    if(widget_name && widgets.custom) {
-        if(!widgets.custom[widget_name]) {
-            return 'missing-custom-' + widget_name;
-        }
-        Widget = widgets.custom[widget_name];
-    } else if(type && widgets.types) {
-        if(!widgets.types[type]) {
-            return 'missing-type-' + type;
-        }
-        Widget = widgets.types[type];
-    }
-
-    let required = List([]);
-    if(parentSchema) {
-        let tmp_required = parentSchema.get('required');
-        if(tmp_required) {
-            required = tmp_required;
-        }
-    }
-
-    // todo: make it easy to to extend these properties with a widget
-    return Widget ? <DumpWidgetRenderer
-        Widget={Widget}// passed to FinalWidgetRenderer and WidgetStackRenderer
-        widgetStack={widgetStack}// passed only to WidgetStackRenderer
-
-        // all others are getting pushed to Widget
-        t={t}
-        {...props}
-        showValidity={showValidity || showValidityOverride}
-        validity={validity}
-        onValidity={onValidity}
-        value={store.getIn(storeKeys)}
-        ownKey={storeKeys.get(storeKeys.count() - 1)}
-        storeKeys={storeKeys}
-        setData={setData}
-        schema={schema}
-        required={required}
-        parentSchema={parentSchema}
-    /> : null;
-};
-
-/**
- * decided if a widget or another group should be rendered, a group is any object without a widget
- */
-let DumpSwitchingRenderer = ({schema, storeKeys, level, groupRenderer: GroupRenderer, dependencies: ownDependencies, ...props}) => {
-    const type = schema.get('type');
-    const widget = schema.get('widget');
-    const properties = schema.get('properties');
-    const dependencies = schema.get('dependencies');
-
-    // it is important here that props get destructed before other props, e.g. `parentSchema` may be overwritten
-    return type !== 'object' || widget ?
-        <SchemaWidgetRenderer
-            {...props}
-            schema={schema} storeKeys={storeKeys} level={level}
-            dependencies={ownDependencies ? ownDependencies : undefined}
-        /> :
-        <GroupRenderer level={level} schema={schema}>
-            {properties ? properties.map((childSchema, childKey) =>
-                <SchemaEditorRenderer
-                    key={childKey}
-                    {...props}
-                    schema={childSchema} parentSchema={schema}
-                    dependencies={dependencies ? dependencies.get(childKey) : undefined}
-                    storeKeys={storeKeys.push(childKey)} level={level + 1}
-                />
-            ).valueSeq() : null}
-        </GroupRenderer>
-};
-DumpSwitchingRenderer = React.memo(DumpSwitchingRenderer);
-
-const SchemaEditorRenderer = ({schema, storeKeys = undefined, level = 0, ...props}) => {
-    const {widgets} = useSchemaEditor();
-
-    if(!storeKeys) {
-        storeKeys = List();
-    }
-
-    const {GroupRenderer} = widgets;
-    if(!GroupRenderer) {
-        console.error('Widget GroupRenderer not existing');
-        return null;
-    }
-
-    return schema ?
-        // split widget/native-types from pure-object
-        <DumpSwitchingRenderer
-            {...props}
-            schema={schema} storeKeys={storeKeys} level={level}
-            groupRenderer={GroupRenderer}
-        />
-        : null;
-};
+import {SchemaEditorRenderer} from "./EditorGroup";
 
 /**
  * @type {React.NamedExoticComponent<{rootRenderer?: *, schema?: *}>}
@@ -224,4 +72,4 @@ let SchemaEditor = ({children, ...props}) => (
 SchemaEditor = React.memo(SchemaEditor);
 
 
-export {SchemaEditor, NestedSchemaEditor, NextPluginRenderer};
+export {SchemaEditor, NestedSchemaEditor};
